% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\name{AddSincast}
\alias{AddSincast}
\alias{AddSincast,Seurat-method}
\alias{Sincast,}
\alias{SincastAssays,}
\alias{Seurat}
\title{Add a \code{SincastAssay} object to Seurat's \code{misc} slots}
\usage{
AddSincast(object, ...)

\S4method{AddSincast}{Seurat}(object)
}
\arguments{
\item{object}{A \code{Seurat} object}
}
\value{
An updated \code{Seurat} object with an additional \code{SincastAssays} object in the \code{misc} slot.
}
\description{
To be added.
}
\seealso{
Other SincastAssays related methods: 
\code{\link{SincastAssays-class}}
}
\concept{SincastAssays related methods}
