% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SincastObject.R
\name{AddSincastObject}
\alias{AddSincastObject}
\alias{Sincast,}
\alias{SincastAssays,}
\alias{Seurat}
\alias{AddSincastObject,Seurat-method}
\title{Add a \code{Sincast} object to \code{Seurat}'s \code{misc} slot.}
\usage{
AddSincastObject(object, replace = FALSE, ...)

\S4method{AddSincastObject}{Seurat}(object, replace = FALSE, ...)
}
\arguments{
\item{object}{A \code{Seurat} object.}

\item{replace}{Logical; if \code{TRUE}, replace the existing \code{Sincast} object in the \code{misc} slot.}
}
\value{
An updated \code{Seurat} object with an additional \code{Sincast} object in the \code{misc} slot.
}
\description{
To be added.
}
\seealso{
Other Sincast related methods: 
\code{\link{CreateSincastObject}()},
\code{\link{Sincast-class}},
\code{\link{SincastSeurat-class}},
\code{\link{SincastToken-class}},
\code{\link{as.SincastSeurat}()}
}
\concept{Sincast related methods}
