% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SincastObject.R
\name{as.Sincast}
\alias{as.Sincast}
\alias{Sincast,}
\alias{SincastObject,}
\alias{Seurat}
\alias{as.Sincast,Seurat-method}
\alias{as.Sincast,Sincast-method}
\title{Convert a \code{Seurat} object to a \code{Sincast} object.}
\usage{
as.Sincast(object, ...)

\S4method{as.Sincast}{Seurat}(object, ...)

\S4method{as.Sincast}{Sincast}(object, ...)
}
\arguments{
\item{object}{A \code{Seurat} object.}
}
\value{
A \code{Sincast} object.
}
\description{
To be added.
}
\seealso{
Other SincastObject related methods: 
\code{\link{CheckSincastObject}()}
}
\concept{SincastObject related methods}
