% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SincastSeurat.R
\name{as.SincastSeurat}
\alias{as.SincastSeurat}
\alias{Sincast,}
\alias{SincastAssays,}
\alias{Seurat}
\title{Convert a \code{Seurat} object to a \code{SincastSeurat} object.}
\usage{
as.SincastSeurat(object, ...)
}
\arguments{
\item{object}{A \code{Seurat} or a \code{SincastSeurat} object.}
}
\value{
A \code{SincastSeurat} object.
}
\description{
To be added.
}
\seealso{
Other Sincast related methods: 
\code{\link{AddSincastObject}()},
\code{\link{CreateSincastObject}()},
\code{\link{Sincast-class}},
\code{\link{SincastSeurat-class}},
\code{\link{SincastToken-class}}
}
\concept{Sincast related methods}
