% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SincastAtlas.R
\name{AtlasPlot}
\alias{AtlasPlot}
\title{Plot \code{Sincast} atlas.}
\usage{
AtlasPlot(
  object,
  atlas = NULL,
  dims = 1:3,
  cells = NULL,
  color.by = "ident",
  colors = NULL,
  anno.by = NULL,
  ...
)

\S4method{AtlasPlot}{Seurat}(
  object,
  atlas = NULL,
  dims = 1:3,
  cells = NULL,
  color.by = "ident",
  colors = NULL,
  anno.by = NULL,
  ...
)

\S4method{AtlasPlot}{Sincast}(
  object,
  atlas = NULL,
  dims = 1:3,
  cells = NULL,
  color.by = "ident",
  colors = NULL,
  anno.by = NULL,
  ...
)
}
\arguments{
\item{object}{A \code{Sincast} atlas, which suppose to be a \code{Seurat} object
generated by \code{\link[Sincast]{BuildSincastAtlas}}.}

\item{atlas}{Either one of the "pseudobulk", "imputation" or "original" indicating the
type of atlas to be plotted Default is the active \code{Sincast} atlas (Can be check by calling the \code{Sincast} object).}

\item{dims}{Dimensions to plot, must be a three-length numeric vector specifying x-, y- and z-dimensions.}

\item{cells}{Cells to plot.}

\item{color.by}{Color cells by which feature or meta.data attribute.}

\item{colors}{Color Scheme of col.by. Should be a named vector using color
codes as values and labels of \code{color.by} as names.}

\item{anno.by}{Additional annotation of cells.}
}
\value{
A \code{plotly} object illustrating 3d principal components.
}
\description{
To be added.
}
\seealso{
Other Sincast plot methods: 
\code{\link{ImputationPlot}()}
}
\concept{Sincast plot methods}
