% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SincastAssays.R
\name{CheckSincastAssays}
\alias{CheckSincastAssays}
\alias{Sincast,}
\alias{SincastAssays}
\alias{CheckSincastAssays,SincastAssays-method}
\alias{CheckSincastAssays,Sincast-method}
\title{Check the validity of the \code{SincastAssays} object of \code{Sincast}.}
\usage{
CheckSincastAssays(object, test = TRUE, slient = FALSE, ...)

\S4method{CheckSincastAssays}{SincastAssays}(object, test = TRUE, silent = FALSE, ...)

\S4method{CheckSincastAssays}{Sincast}(object, test = TRUE, silent = FALSE, ...)
}
\arguments{
\item{object}{A \code{SincastAssays} object.}

\item{test}{Logical; if TRUE (the default) and validity fails, the function returns a
vector of strings describing the problems. If test is FALSE validity failure generates an error.}

\item{silent}{Logical; if TRUE, suppress all messages.}
}
\value{
A vector of strings describing the problems.
}
\description{
To be added.
}
\seealso{
Other SincastAssays related methods: 
\code{\link{CleanSincastAssays}()},
\code{\link{GetSincastAssays}()}
}
\concept{SincastAssays related methods}
