% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SincastAtlas.R
\name{CheckSincastAtlas}
\alias{CheckSincastAtlas}
\alias{Sincast,}
\alias{SincastAtlas}
\alias{CheckSincastAtlas,SincastAtlas-method}
\alias{CheckSincastAtlas,Sincast-method}
\title{Check the validity of the \code{SincastAtlas} object of \code{Sincast}.}
\usage{
CheckSincastAtlas(object, test = TRUE, slient = FALSE, ...)

\S4method{CheckSincastAtlas}{SincastAtlas}(object, test = TRUE, silent = FALSE, ...)

\S4method{CheckSincastAtlas}{Sincast}(object, test = TRUE, silent = FALSE, ...)
}
\arguments{
\item{object}{A \code{SincastAtlas} object.}

\item{test}{Logical; if TRUE (the default) and validity fails, the function returns a
vector of strings describing the problems. If test is FALSE validity failure generates an error.}

\item{silent}{Logical; if TRUE, suppress all messages.}
}
\value{
A vector of strings describing the problems.
}
\description{
To be added.
}
\seealso{
Other SincastAtlas related methods: 
\code{\link{CleanSincastAtlas}()},
\code{\link{GetSincastAtlas}()}
}
\concept{SincastAtlas related methods}
