% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SincastObject.R
\name{CheckSincastObject}
\alias{CheckSincastObject}
\alias{CheckSincastObject,Sincast-method}
\title{Check the validity of the \code{Sincast} object.}
\usage{
CheckSincastObject(object, test = TRUE, complete = TRUE, slient = FALSE, ...)

\S4method{CheckSincastObject}{Sincast}(object, test = TRUE, complete = TRUE, silent = FALSE, ...)
}
\arguments{
\item{object}{A \code{Sincast} object.}

\item{test}{Logical; if TRUE (the default) and validity fails, the function returns a
vector of strings describing the problems. If test is FALSE validity failure generates an error.}

\item{complete}{Logical; if TRUE, call validity check for each slot of the
existing \code{Sincast} object.}

\item{silent}{Logical; if TRUE, suppress all messages.}
}
\value{
A vector of strings describing the problems.
}
\description{
To be added.
}
\seealso{
Other SincastObject related methods: 
\code{\link{as.Sincast}()}
}
\concept{SincastObject related methods}
