% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SincastAssays.R
\name{CleanSincastAssays}
\alias{CleanSincastAssays}
\alias{Sincast,}
\alias{SincastAssays,}
\alias{Seurat}
\alias{CleanSincastAssays,Seurat-method}
\title{Remove unwanted or invalid elements (i.e, non-\code{Seurat} object) in \code{Sincast} assays
\code{pseudobulk} and \code{imputation} embedded in \code{Seurat::MISC(object, slot = "SincastAssays")}.}
\usage{
CleanSincastAssays(
  object,
  remove.id.pseudobulk = NULL,
  remove.id.imputation = NULL,
  clean.up = c("None", "All", "pseudobulk", "imputation"),
  remove.invalid = TRUE,
  ...
)

\S4method{CleanSincastAssays}{Seurat}(
  object,
  remove.id.pseudobulk = NULL,
  remove.id.imputation = NULL,
  clean.up = c("None", "All", "pseudobulk", "imputation"),
  remove.invalid = TRUE
)
}
\arguments{
\item{object}{A \code{Seurat} object with an additional \code{SincastAssays} object in the \code{misc} slot.}

\item{remove.id.pseudobulk}{A vector of ids of the elements to be removed from the \code{pseduobulk} assay.}

\item{remove.id.imputation}{A vector of ids of the elements to be removed from the \code{Imputation} assay.}

\item{clean.up}{Either "All", "pseudobulk" or "imputation" indicating whether to clean up
the \code{SincastAssays} object or clean up a specific \code{Sincast} assay as proposed.}

\item{remove.invalid}{Logical. Whether to check the validity of the existing \code{SincastAssays} object, and
remove the invalid elements (e.g, non-\code{Seurat} object) identified in \code{Sincast} assays.}
}
\value{
An updated \code{Seurat} object with an additional \code{SincastAssays} object in the \code{misc} slot.
}
\description{
To be added.
}
\seealso{
Other SincastAssays related methods: 
\code{\link{CreateSincastAssays}()},
\code{\link{SincastAggregate}()},
\code{\link{SincastAssays-class}}
}
\concept{SincastAssays related methods}
