% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SincastAssays.R
\name{CreateSincastAssays}
\alias{CreateSincastAssays}
\alias{Sincast,}
\alias{SincastAssays,}
\alias{Seurat}
\alias{CreateSincastAssays,Seurat-method}
\title{Add a \code{SincastAssays} object to \code{Seurat}'s \code{misc} slot.}
\usage{
CreateSincastAssays(object, replace = FALSE, ...)

\S4method{CreateSincastAssays}{Seurat}(object, replace = FALSE)
}
\arguments{
\item{object}{A \code{Seurat} object.}

\item{replace}{Whether to replace the existing \code{SincastAssays} object.}
}
\value{
An updated \code{Seurat} object with an additional \code{SincastAssays} object in the \code{misc} slot.
}
\description{
To be added.
}
\seealso{
Other SincastAssays related methods: 
\code{\link{CleanSincastAssays}()},
\code{\link{SincastAggregate}()},
\code{\link{SincastAssays-class}}
}
\concept{SincastAssays related methods}
