% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SincastObject.R
\name{CreateSincastObject}
\alias{CreateSincastObject}
\alias{Sincast,}
\alias{SincastAssays}
\title{Create an empty \code{Sincast} object.}
\usage{
CreateSincastObject(
  by = "CreateSincastObject",
  command = deparse(match.call())
)
}
\value{
An empty \code{Sincast} object.
}
\description{
To be added.
}
\seealso{
Other Sincast related methods: 
\code{\link{as.SincastSeurat}()}
}
\concept{Sincast related methods}
