% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExternalFunctions.R
\name{ExternalFunctions}
\alias{ExternalFunctions}
\alias{\%>\%}
\title{External functions}
\usage{
lhs \%>\% rhs
}
\description{
These functions are imported from other packages: \code{\link[magrittr]{\%>\%}}.
}
