% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SincastAssays.R
\name{GetSincastAssays}
\alias{GetSincastAssays}
\alias{Sincast,}
\alias{SincastAssays}
\alias{GetSincastAssays,Sincast-method}
\alias{GetSincastAssays<-,Sincast-method}
\title{Extract the \code{SincastAssays} object from a \code{Sincast} object.}
\usage{
GetSincastAssays(
  object,
  assay = c("all", "original", "pseudobulk", "imputation"),
  ...
)

\S4method{GetSincastAssays}{Sincast}(
  object,
  assay = c("all", "original", "pseudobulk", "imputation"),
  ...
)

GetSincastAssays(
  object,
  assay = c("all", "original", "pseudobulk", "imputation"),
  ...
) <- value

\S4method{GetSincastAssays}{Sincast}(
  object,
  assay = c("all", "original", "pseudobulk", "imputation"),
  ...
) <- value
}
\arguments{
\item{object}{A \code{Sincast} object.}

\item{assay}{Either \code{"all"}, \code{"original"}, \code{"pseudobulk"} or
\code{"imputation"} indicating which specific \code{Sincast} assay to modify.}
}
\value{
Depending on the \code{assay} argument, can be either type of a \code{Sicast} assay,
or a \code{SincastAssays} object containing all the types.

A \code{Sincast} object with updated \code{Sincast} assays.
}
\description{
To be added.

To be added.
}
\seealso{
Other SincastAssays related methods: 
\code{\link{CheckSincastAssays}()},
\code{\link{CleanSincastAssays}()}

Other SincastAssays related methods: 
\code{\link{CheckSincastAssays}()},
\code{\link{CleanSincastAssays}()}
}
\concept{SincastAssays related methods}
