% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SincastAtlas.R
\name{GetSincastAtlas}
\alias{GetSincastAtlas}
\alias{GetSincastAtlas,Sincast-method}
\alias{GetSincastAtlas<-,Sincast-method}
\title{Extract the \code{SincastAtlas} object from a \code{Sincast} object.}
\usage{
GetSincastAtlas(
  object,
  atlas = c("all", "original", "pseudobulk", "imputation"),
  ...
)

\S4method{GetSincastAtlas}{Sincast}(
  object,
  atlas = c("all", "original", "pseudobulk", "imputation"),
  ...
)

GetSincastAtlas(
  object,
  atlas = c("all", "original", "pseudobulk", "imputation"),
  ...
) <- value

\S4method{GetSincastAtlas}{Sincast}(
  object,
  atlas = c("all", "original", "pseudobulk", "imputation"),
  ...
) <- value
}
\arguments{
\item{object}{A \code{Sincast} object.}

\item{atlas}{Either \code{"all"}, \code{"original"}, \code{"pseudobulk"} or
\code{"imputation"} indicating which specific \code{Sincast} atlas to modify.}
}
\value{
Depending on the \code{atlas} argument, can be either type of a \code{Sicast} atlas,
or a \code{SincastAtlas} object containing all the types.

A \code{Sincast} object with updated \code{Sincast} atlas.
}
\description{
To be added.

To be added.
}
\seealso{
Other SincastAtlas related methods: 
\code{\link{BuildSincastAtlas}()},
\code{\link{CheckSincastAtlas}()},
\code{\link{CleanSincastAtlas}()}

Other SincastAtlas related methods: 
\code{\link{BuildSincastAtlas}()},
\code{\link{CheckSincastAtlas}()},
\code{\link{CleanSincastAtlas}()}
}
\concept{SincastAtlas related methods}
