% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SincastObject.R
\name{GetSincastObject}
\alias{GetSincastObject}
\alias{Sincast,}
\alias{SincastObject,}
\alias{Seurat}
\alias{GetSincastObject,Seurat-method}
\alias{GetSincastObject<-}
\alias{GetSincastObject<-,Seurat-method}
\title{Extract the \code{Sincast} object from \code{Seurat}'s \code{misc} slot.}
\usage{
GetSincastObject(object, ...)

\S4method{GetSincastObject}{Seurat}(object, ...)

GetSincastObject(object, ...) <- value

\S4method{GetSincastObject}{Seurat}(object, ...) <- value
}
\arguments{
\item{object}{A \code{Seurat} object.}
}
\value{
A \code{Sincast} object.

A \code{Seurat} object with updated \code{Sincast} object
}
\description{
To be added.

To be added.
}
\seealso{
Other SincastObject related methods: 
\code{\link{CheckSincastObject}()},
\code{\link{CreateSincastObject}()}

Other SincastObject related methods: 
\code{\link{CheckSincastObject}()},
\code{\link{CreateSincastObject}()}
}
\concept{SincastObject related methods}
