% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SincastImpute.R
\name{ImputationPlot}
\alias{ImputationPlot}
\title{Plot the diffusion components computed on the diffusion operator used for \code{Sincast}
imputation}
\usage{
ImputationPlot(
  object,
  dims = 1:3,
  cells = NULL,
  color.by = "ident",
  colors = NULL,
  anno.by = NULL,
  ...
)

\S4method{ImputationPlot}{Seurat}(
  object,
  dims = 1:3,
  cells = NULL,
  color.by = "ident",
  colors = NULL,
  anno.by = NULL,
  ...
)

\S4method{ImputationPlot}{Sincast}(
  object,
  dims = 1:3,
  cells = NULL,
  color.by = "ident",
  colors = NULL,
  anno.by = NULL,
  ...
)
}
\arguments{
\item{object}{\code{Sincast}'s imputation assay, which suppose to be a \code{Seurat} object
generated by \code{\link[Sincast]{SincastImpute}}.}

\item{dims}{Dimensions to plot, must be a three-length numeric vector specifying x-, y- and z-dimensions.}

\item{cells}{Cells to plot.}

\item{color.by}{Color query cells by which feature or meta.data attribute.}

\item{colors}{Color Scheme of col.by. Should be a named vector using color
codes as values and labels of \code{color.by} as names.}

\item{anno.by}{Additional annotation of query cells.}
}
\value{
A \code{plotly} object.
}
\description{
To be added.
}
\seealso{
\code{\link[Sincast]{SincastAggregate}()}
}
\concept{Sincast plot methods}
