% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Others.R
\name{IsLogicalVector}
\alias{IsLogicalVector}
\title{Check if a \code{R} variable is a logical vector.}
\usage{
IsLogicalVector(x, n = NULL, in.which = NULL)
}
\arguments{
\item{x}{a \code{R} code}

\item{n}{Check if \code{x} is a logical vector of length \code{n}.}

\item{in.which}{A logical vector. Check if \code{x} is a subset of \code{in.which}.}
}
\value{
Logical indicates the test result.
}
\description{
To be added.
}
