% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Others.R
\name{IsNumericVector}
\alias{IsNumericVector}
\title{Check if a \code{R} variable is a numerical vector.}
\usage{
IsNumericVector(x, n = NULL, in.which = NULL)
}
\arguments{
\item{x}{a \code{R} code}

\item{n}{Check if \code{x} is a numerical vector of length \code{n}.}

\item{in.which}{A numerical vector. Check if \code{x} is a subset of \code{in.which}.}
}
\value{
Logical indicates the test result.
}
\description{
To be added.
}
