% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SincastImpute.R
\name{LowRankApprox}
\alias{LowRankApprox}
\title{Low rank approximation.}
\usage{
LowRankApprox(
  x,
  k = NULL,
  estimate.rank = TRUE,
  do.rsvd = TRUE,
  seed = 521626,
  q = 0.9,
  ...
)
}
\arguments{
\item{x}{a non-negative feature-by-cell matrix.}

\item{k}{Number of singular values requested for single value decomposition. If \code{estimate.rank} is set to TRUE, default is 100, otherwise 50.}

\item{estimate.rank}{Logical; If TRUE, the rank of \code{x} will be estimated by the approach proposed by Linderman et al (2022),
which will be less or equal to \code{k}. Otherwise the rank will be set to \code{k}. Also note that if TRUE, \code{k} should be a numeric larger than 100.}

\item{do.rsvd}{Logical; If TRUE, perform random single value decomposition.}

\item{seed}{The random seed for random SVD if do.rsvd = TRUE.}

\item{q}{If a svd-approximation of a feature contains negative values, calculate the qth smallest negative value. Any imputed expression that are lower
than the absolute of this value is set to zero.}
}
\value{
A low rank approximation of \code{x}
}
\description{
Perform the scRNA-seq low rank approximation method proposed by
\href{https://www.nature.com/articles/s41467-021-27729-z}{George C. Linderman (2022, nat communication)}.
The major contribution of the method is that it can perform imputation while preserve biological zeros.
}
