% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Others.R
\name{RankTrans}
\alias{RankTrans}
\title{Perform rank transformation}
\usage{
RankTrans(x)
}
\arguments{
\item{x}{a sample by feature matrix.}
}
\value{
A rank transformed data matrix.
}
\description{
To be added.
}
