% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassDefinition.R
\docType{class}
\name{Sincast-class}
\alias{Sincast-class}
\alias{Sincast}
\alias{Sincast,}
\alias{SincastAssays}
\title{An S4 class of \code{Sincast} object.}
\description{
To be added.
}
\section{Slots}{

\describe{
\item{\code{SincastAssays}}{A S4 class to store \code{Sincast} aggregation or imputation results.}

\item{\code{SincastToken}}{A S4 class to mark and time stamp the \code{Sincast} object.}
}}

\seealso{
Other Sincast related methods: 
\code{\link{AddSincastObject}()},
\code{\link{CreateSincastObject}()},
\code{\link{SincastToken-class}}
}
\concept{Sincast related methods}
