% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SincastAggregate.R
\name{SincastAggregate}
\alias{SincastAggregate}
\alias{Sincast,}
\alias{SincastAssays,}
\alias{Seurat}
\alias{SincastAggregate,Seurat-method}
\title{Aggregate single cells to generate pseudobulk samples.}
\usage{
SincastAggregate(
  object,
  assay = NULL,
  layer = "counts",
  features = NULL,
  group.by = "ident",
  sample.method = c("equal.size", "equal.prop"),
  aggregate.method = c("addition", "average"),
  n.pool = 15,
  size.factor = 1,
  pool.factor = NULL,
  sep = ".",
  remove.invalid = TRUE,
  ...
)

\S4method{SincastAggregate}{Seurat}(
  object,
  assay = NULL,
  layer = "counts",
  features = NULL,
  group.by = "ident",
  sample.method = c("equal.size", "equal.prop"),
  aggregate.method = c("addition", "average"),
  n.pool = 15,
  size.factor = 1,
  pool.factor = NULL,
  sep = ".",
  remove.invalid = TRUE,
  ...
)
}
\arguments{
\item{object}{A \code{Seurat} object}

\item{assay}{To be added.}

\item{layer}{To be added.}

\item{features}{To be added.}

\item{group.by}{To be added.}

\item{sample.method}{To be added.}

\item{n.pool}{To be added.}

\item{size.factor}{To be added.}

\item{pool.factor}{To be added.}

\item{sep}{To be added.}

\item{remove.invalid}{Logical. Whether to check the validity of the existing \code{SincastAssays} object, and
remove the invalid elements (e.g, non-\code{Seurat} object) identified in \code{Sincast} assays.}
}
\value{
An updated \code{Seurat} object with an additional \code{SincastAssays}
object in the \code{misc} slot. Aggregated pseudobulk samples are stored as an appended element
of the Sincast assay \code{pseudobulk} embedded in \code{Seurat::MISC(object, slot = 'SincastAssays')}.
}
\description{
To be added.
}
\seealso{
Other SincastAssays related methods: 
\code{\link{CleanSincastAssays}()},
\code{\link{CreateSincastAssays}()},
\code{\link{SincastAssays-class}}
}
\concept{SincastAssays related methods}
