% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassDefinition.R
\docType{class}
\name{SincastAssays-class}
\alias{SincastAssays-class}
\alias{SincastAssays}
\alias{Sincast,}
\title{A S4 class to store \code{Sincast} aggregation or imputation results.}
\description{
To be added.
}
\section{Slots}{

\describe{
\item{\code{original}}{A \code{Seurat} object storing the original single cell data.}

\item{\code{pseudobulk}}{A \code{Seurat} object storing aggregated pseudobulk.}

\item{\code{imputation}}{A \code{Seurat} object storing imputed single cells.}
}}

\seealso{
Other Sincast classes: 
\code{\link{Sincast-class}},
\code{\link{SincastSummary-class}},
\code{\link{SincastToken-class}}
}
\concept{Sincast classes}
