% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassDefinition.R
\docType{class}
\name{SincastAtlas-class}
\alias{SincastAtlas-class}
\alias{SincastAtlas}
\alias{Sincast,}
\title{A S4 class to store \code{Sincast} pseudobulk or imputation atlas}
\description{
To be added.
}
\section{Slots}{

\describe{
\item{\code{original}}{A \code{Seurat} object storing the atlas constructed on the original data.}

\item{\code{pseudobulk}}{A \code{Seurat} object storing the pseudobulk atlas.}

\item{\code{imputation}}{A \code{Seurat} object storing the imputation atlas.}
}}

\seealso{
Other Sincast classes: 
\code{\link{Sincast-class}},
\code{\link{SincastAssays-class}},
\code{\link{SincastSummary-class}},
\code{\link{SincastToken-class}}
}
\concept{Sincast classes}
