% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassDefinition.R
\docType{class}
\name{SincastSeurat-class}
\alias{SincastSeurat-class}
\alias{SincastSeurat}
\alias{Sincast,}
\alias{SincastAssays,}
\alias{Seurat}
\title{An S4 class of \code{SincastSeurat} object, extended based on \code{Seurat}.}
\description{
To be added.
}
\section{Slots}{

\describe{
\item{\code{Sincast}}{An S4 class of \code{Sincast} object.}
}}

\seealso{
Other Sincast related methods: 
\code{\link{AddSincastObject}()},
\code{\link{CreateSincastObject}()},
\code{\link{Sincast-class}},
\code{\link{SincastToken-class}},
\code{\link{as.SincastSeurat}()}
}
\concept{Sincast related methods}
