% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassDefinition.R
\docType{class}
\name{SincastSummary-class}
\alias{SincastSummary-class}
\alias{SincastSummary}
\alias{Sincast}
\alias{SincastSummary-validity}
\alias{Sincast,}
\title{An S4 class of \code{SincastSummary} object.}
\description{
To be added.

Validation of \code{SincastSummary} objects is handled by \code{\link[methods]{validObject}}.
}
\section{Slots}{

\describe{
\item{\code{summary}}{A \code{data.frame} summarize \code{Sincast} results.}
}}

\seealso{
Other Sincast classes: 
\code{\link{Sincast-class}},
\code{\link{SincastAssays-class}},
\code{\link{SincastToken-class}}
}
\concept{Sincast classes}
