% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SincastSeurat.R
\name{as.SincastSeurat}
\alias{as.SincastSeurat}
\alias{Sincast,}
\alias{Seurat}
\title{Convert a \code{Seurat} object to a \code{SincastSeurat} object.}
\usage{
as.SincastSeurat(object, ...)
}
\arguments{
\item{object}{A \code{Seurat} or a \code{SincastSeurat} object.}
}
\value{
A \code{SincastSeurat} object.
}
\description{
To be added.
}
\concept{SincastSeurat related methods}
