% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SincastAtlas.R
\name{BuildSincastAtlas}
\alias{BuildSincastAtlas}
\alias{BuildSincastAtlas,Sincast-method}
\title{Build a \code{Sincast} atlas, which is a \code{Seurat} object with a rank-transformed reference
data, as well a pca dimensional reduction performed on the data.}
\usage{
BuildSincastAtlas(
  object,
  Sincast.assay = NULL,
  Seurat.assay = NULL,
  layer = NULL,
  reference.features = NULL,
  query.features = NULL,
  ScaleData.control = lits(),
  RunPCA.control = list(),
  ...
)

\S4method{BuildSincastAtlas}{Sincast}(
  object,
  Sincast.assay = NULL,
  Seurat.assay = NULL,
  layer = NULL,
  reference.features = NULL,
  query.features = NULL,
  ScaleData.control = lits(),
  RunPCA.control = list(),
  ...
)
}
\arguments{
\item{object}{A \code{Sincast} object.}

\item{Sincast.assay}{description}

\item{Seurat.assay}{description}

\item{layer}{description}

\item{reference.features}{description}

\item{query.features}{description}

\item{ScaleData.control}{description}

\item{RunPCA.control}{description}
}
\value{
A \code{Sincast} object with updated \code{Sincast} atlas.
}
\description{
To be added.
}
\seealso{
Other SincastAtlas related methods: 
\code{\link{CheckSincastAtlas}()},
\code{\link{CleanSincastAtlas}()},
\code{\link{GetSincastAtlas}()}
}
\concept{SincastAtlas related methods}
