% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SincastObject.R
\name{CheckSincastObject}
\alias{CheckSincastObject}
\alias{Sincast,}
\alias{SincastAssays,}
\alias{Seurat}
\alias{CheckSincastObject,Seurat-method}
\title{Check the validity of the \code{Sincast} object in \code{Seurat}'s \code{misc} slot.}
\usage{
CheckSincastObject(object, test = TRUE, complete = TRUE, ...)

\S4method{CheckSincastObject}{Seurat}(object, test = TRUE, complete = TRUE, ...)
}
\arguments{
\item{object}{A \code{Seurat} object.}

\item{test}{Logical; if TRUE (the default) and validity fails, the function returns a
vector of strings describing the problems. If test is FALSE validity failure generates an error.}

\item{complete}{Logical; if TRUE, call validity check for each slot of the
existing \code{Sincast} object.}
}
\value{
Logical indicates whether \code{Seurat}'s \code{misc} slot
contains a valid \code{Sincast} object.
}
\description{
To be added.
}
\seealso{
Other SincastObject related methods: 
\code{\link{GetSincastObject}()}
}
\concept{SincastObject related methods}
