% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SincastAssays.R
\name{CleanSincastAssays}
\alias{CleanSincastAssays}
\alias{Sincast,}
\alias{SincastAssays,}
\alias{Seurat}
\alias{CleanSincastAssays,Seurat-method}
\title{Remove unwanted or invalid elements (i.e, non-\code{Seurat} object) in \code{Sincast} assays
\code{pseudobulk} and \code{imputation} embedded in \code{Sincast} appended \code{Seurat} object.}
\usage{
CleanSincastAssays(
  object,
  remove.id.pseudobulk = NULL,
  remove.id.imputation = NULL,
  clean.up = c("none", "all", "pseudobulk", "imputation"),
  remove.invalid = TRUE,
  ...
)

\S4method{CleanSincastAssays}{Seurat}(
  object,
  remove.id.pseudobulk = NULL,
  remove.id.imputation = NULL,
  clean.up = c("none", "all", "pseudobulk", "imputation"),
  remove.invalid = TRUE,
  ...
)
}
\arguments{
\item{object}{A \code{Sincast} appended \code{Seurat} object.}

\item{clean.up}{Either \code{"None"}, \code{"All"}, \code{"pseudobulk"} or \code{"imputation"} indicating whether to clean up
the \code{SincastAssays} object or clean up a specific \code{Sincast} assay as proposed.}

\item{remove.invalid}{Logical; if TRUE, check the validity of the existing \code{Sincast} assays and
remove invalid ones (e.g, \code{Seurat} objects not generated by \code{Sincast} functions).}
}
\value{
A \code{Seurat} object with updated \code{Sincast} assays.
}
\description{
To be added.
}
\seealso{
Other SincastAssays related methods: 
\code{\link{GetSincastAssays}()}
}
\concept{SincastAssays related methods}
