% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SincastAssays.R
\name{CleanSincastAssays}
\alias{CleanSincastAssays}
\alias{CleanSincastAssays,SincastAssays-method}
\alias{CleanSincastAssays,Sincast-method}
\title{Remove unwanted or invalid (i.e, non-\code{Seurat} object) in \code{Sincast} assays
(\code{pseudobulk} and \code{imputation}) embedded in \code{Sincast}.}
\usage{
CleanSincastAssays(
  object,
  clean.up = c("none", "all", "pseudobulk", "imputation"),
  remove.invalid = TRUE,
  ...
)

\S4method{CleanSincastAssays}{SincastAssays}(
  object,
  clean.up = c("none", "all", "pseudobulk", "imputation"),
  remove.invalid = TRUE,
  ...
)

\S4method{CleanSincastAssays}{Sincast}(
  object,
  clean.up = c("none", "all", "pseudobulk", "imputation"),
  remove.invalid = TRUE,
  ...
)
}
\arguments{
\item{object}{A \code{Sincast} object.}

\item{clean.up}{Either \code{"None"}, \code{"All"}, \code{"pseudobulk"} or \code{"imputation"} indicating whether to clean up
the \code{SincastAssays} object or clean up a specific \code{Sincast} assay as proposed.}

\item{remove.invalid}{Logical; if TRUE, check the validity of the existing \code{Sincast} assays and
remove invalid ones (e.g, \code{Seurat} objects not generated by \code{Sincast} functions).}
}
\value{
A \code{Sincast} object with updated \code{Sincast} assays.
}
\description{
To be added.
}
\seealso{
Other SincastAssays related methods: 
\code{\link{CheckSincastAssays}()},
\code{\link{GetSincastAssays}()}
}
\concept{SincastAssays related methods}
