% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SincastAssays.R
\name{CreateSincastAssays}
\alias{CreateSincastAssays}
\alias{Sincast,}
\alias{SincastAssays,}
\alias{Seurat}
\alias{CreateSincastAssays,Seurat-method}
\title{Add a \code{SincastAssays} object to \code{Seurat}'s \code{misc} slot.}
\usage{
CreateSincastAssays(object, remove.invalid = TRUE, ...)

\S4method{CreateSincastAssays}{Seurat}(object, remove.invalid = TRUE)
}
\arguments{
\item{object}{A \code{Seurat} object.}

\item{remove.invalid}{Logical. Whether to check the validity of the existing \code{Sincast} assays and
remove invalid ones (e.g, \code{Seurat} objects not generated by \code{Sincas}t functions).}
}
\value{
An updated \code{Seurat} object with an additional \code{SincastAssays} object in the \code{misc} slot.
}
\description{
To be added.
}
\seealso{
Other SincastAssays related methods: 
\code{\link{CleanSincastAssays}()},
\code{\link{SincastAggregate}()},
\code{\link{SincastAssays-class}}
}
\concept{SincastAssays related methods}
