% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SincastObject.R
\name{CreateSincastObject}
\alias{CreateSincastObject}
\alias{Sincast,}
\alias{SincastObject}
\title{Create an empty \code{Sincast} object.}
\usage{
CreateSincastObject(
  by = "CreateSincastObject",
  command = deparse(match.call())
)
}
\value{
An empty \code{Sincast} object.
}
\description{
To be added.
}
\section{Slots}{

\describe{
\item{\code{by}}{A string recording the function by which \code{CreateSincastObject} was called.}

\item{\code{command}}{A string recording \code{Sincast} command history.}
}}

\seealso{
Other SincastObject related methods: 
\code{\link{CheckSincastObject}()},
\code{\link{GetSincastObject}()}
}
\concept{SincastObject related methods}
