% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SincastAssays.R
\name{GetSincastAssays}
\alias{GetSincastAssays}
\alias{Sincast,}
\alias{SincastAssays,}
\alias{Seurat}
\alias{GetSincastAssays,Seurat-method}
\alias{GetSincastAssays<-}
\alias{GetSincastAssays<-,Seurat-method}
\title{Extract the \code{SincastAssays} object from a \code{Sincast} appended \code{Seurat} object.}
\usage{
GetSincastAssays(object, assay = c("both", "pseudobulk", "imputation"), ...)

\S4method{GetSincastAssays}{Seurat}(object, assay = c("both", "pseudobulk", "imputation"), ...)

GetSincastAssays(object, assay = c("both", "pseudobulk", "imputation"), ...) <- value

\S4method{GetSincastAssays}{Seurat}(object, assay = c("both", "pseudobulk", "imputation"), ...) <- value
}
\arguments{
\item{object}{A \code{Seurat} object.}

\item{assay}{Either \code{"both"}, \code{"pseudobulk"} or \code{"imputation"} indicating
which specific \code{Sincast} assay to be set.}
}
\value{
Depending on the \code{assay} argument, can be either type of a \code{Sicast} assay,
or a \code{SincastAssays} object containing both the types.

A \code{Seurat} object with updated \code{Sincast} assays.
}
\description{
To be added.

To be added.
}
\seealso{
Other SincastAssays related methods: 
\code{\link{CleanSincastAssays}()}

Other SincastAssays related methods: 
\code{\link{CleanSincastAssays}()}
}
\concept{SincastAssays related methods}
