% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SincastImpute.R
\name{LowRankApprox}
\alias{LowRankApprox}
\title{Low rank approximation.}
\usage{
LowRankApprox(x, rank = NULL, do.rsvd = TRUE, seed = 521626, q = 0.9, ...)
}
\arguments{
\item{x}{a non-negative feature-by-cell matrix.}

\item{rank}{The rank of the approximating matrix of \code{x}. If \code{NULL}, it will be determined by the approach proposed by Linderman et al (2022).}

\item{do.rsvd}{Logical; If TRUE, perform random single value decomposition.}

\item{seed}{The random seed for random SVD if do.rsvd = TRUE.}

\item{q}{If a svd-approximation of a feature contains negative values, calculate the qth smallest negative value. Any imputed expression that are lower
than the absolute of this value is set to zero.}
}
\value{
A low rank approximation of \code{x}
}
\description{
Perform the scRNA-seq low rank approximation method proposed by
\href{https://www.nature.com/articles/s41467-021-27729-z}{George C. Linderman (2022, nat communication)}.
The major contribution of the method is that it can perform imputation while preserve biological zeros.
}
