% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassDefinition.R
\docType{class}
\name{Sincast-class}
\alias{Sincast-class}
\alias{Sincast}
\alias{Sincast,}
\alias{SincastAssays}
\title{An S4 class of \code{Sincast} object.}
\description{
To be added.
}
\section{Slots}{

\describe{
\item{\code{SincastAssays}}{A S4 class to store \code{Sincast} aggregation or imputation results.}

\item{\code{Summary}}{A \code{SincastSummary} object.}
}}

\seealso{
Other Sincast classes: 
\code{\link{SincastAssays-class}},
\code{\link{SincastSummary-class}},
\code{\link{SincastToken-class}}
}
\concept{Sincast classes}
