% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SincastAggregate.R
\name{SincastAggregate}
\alias{SincastAggregate}
\alias{Sincast,}
\alias{SincastAssays,}
\alias{Seurat}
\alias{SincastAggregate,Seurat-method}
\title{Aggregate single cells to generate pseudobulk samples.}
\usage{
SincastAggregate(
  object,
  assay = NULL,
  layer = "counts",
  cells = NULL,
  features = NULL,
  group.by = "ident",
  sample.method = c("equal.size", "equal.prop"),
  aggregate.method = c("addition", "average"),
  n.pool = 15,
  size.factor = 1,
  pool.factor = NULL,
  sep = ".",
  replace = c("None", "All", "invalid"),
  ...
)

\S4method{SincastAggregate}{Seurat}(
  object,
  assay = NULL,
  layer = "counts",
  cells = NULL,
  features = NULL,
  group.by = "ident",
  sample.method = c("equal.size", "equal.prop"),
  aggregate.method = c("addition", "average"),
  n.pool = 15,
  size.factor = 1,
  pool.factor = NULL,
  sep = ".",
  replace = FALSE,
  ...
)
}
\arguments{
\item{object}{A \code{Seurat} object}

\item{assay}{To be added.}

\item{layer}{To be added.}

\item{cells}{To be added.}

\item{features}{To be added.}

\item{group.by}{To be added.}

\item{sample.method}{To be added.}

\item{n.pool}{To be added.}

\item{size.factor}{To be added.}

\item{pool.factor}{To be added.}

\item{sep}{To be added.}

\item{replace}{Whether to replace the existing 'pseudobulk' assay.}
}
\value{
A \code{Seurat} object with updated \code{Sincast pseudobulk} assay.
}
\description{
To be added.
}
