% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SincastAggregate.R
\name{SincastAggregate}
\alias{SincastAggregate}
\title{Aggregate single cells to generate pseudobulk samples.}
\usage{
SincastAggregate(
  object,
  assay = NULL,
  layer = "counts",
  features = NULL,
  group.by = "ident",
  sample.method = c("equal.size", "equal.prop"),
  aggregate.method = c("addition", "average"),
  n.pool = 15,
  size.factor = 1,
  pool.factor = NULL,
  sep = "_",
  replace = FALSE,
  ...
)

\S4method{SincastAggregate}{Sincast}(
  object,
  assay = NULL,
  layer = "counts",
  features = NULL,
  group.by = "ident",
  sample.method = c("equal.size", "equal.prop"),
  aggregate.method = c("addition", "average"),
  n.pool = 15,
  size.factor = 1,
  pool.factor = NULL,
  sep = "_",
  replace = FALSE,
  ...
)
}
\arguments{
\item{object}{A \code{Sincast} object.}

\item{assay}{Name of the \code{Seurat} assay aggregation is being run on. Default is the default \code{Seurat} assay.}

\item{layer}{Name of the \code{Seurat} layer aggregation is being run on. Default is the \code{counts} layer.}

\item{features}{Features to analyze. Default is all features in the assay.}

\item{group.by}{To be added.}

\item{sample.method}{To be added.}

\item{n.pool}{To be added.}

\item{size.factor}{To be added.}

\item{pool.factor}{To be added.}

\item{sep}{To be added.}

\item{replace}{Logical; if TRUE, replace the existing \code{pseudobulk} assay.}
}
\value{
A \code{Sincast} object with updated \code{pseudobulk} assay.
}
\description{
Perform \code{Sincast} aggregation.
}
\seealso{
\code{\link[Sincast]{SincastImpute}()}
}
