% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassDefinition.R
\docType{class}
\name{SincastSeurat-class}
\alias{SincastSeurat-class}
\alias{SincastSeurat}
\alias{Sincast,}
\alias{Seurat}
\title{An S4 class of \code{SincastSeurat} object, extended based on \code{Seurat}.}
\description{
Inherit all slots from a \code{Seurat} object (See \code{\link[Seurat]{Seurat-class}}),
plus an additional slot \code{Sincast} storing a \code{Sincast}
object (See \code{\link[Sincast]{Sincast-class}}).
}
\section{Slots}{

\describe{
\item{\code{Sincast}}{An S4 class of \code{Sincast} object.}
}}

\seealso{
[as.SincastSeurat()] for converting a \code{Seurat} object to \code{Sincast}.

Other Sincast classes: 
\code{\link{Sincast-class}},
\code{\link{SincastAssays-class}},
\code{\link{SincastToken-class}}
}
\concept{Sincast classes}
